refactor: move sassdoc to its own helper
To be used by other features
